Add files via upload
This is a modified version of my AnalogWaterSensor file. I added a transistor that turns on an LED. I didn't know how to test the transistor's base threshold (I assume it's 0.7V) since I lack some equipment like a multimeter. I tried using the water sensor's signal pin since it outputs a variable voltage from 0V to 1V (Not sure why it only goes to 1V), but the transistor wouldn't budge for some reason. This program is basically me coping with that fact by finding the voltage from the water sensor myself, and when the voltage reaches 0.7, the Arduino sends a signal to the base of the transistor to turn on. 
TL;DR Transistorception